add food screen

+++ backend/app/Http/Controllers/CategoryController.php
@@ -50,7 +50,9 @@ class CategoryController extends Controller
             $imageName = time() . '_' . $request->category_image->getClientOriginalName();
             $request->category_image->move(public_path('img'), $imageName);
             $oldImagePath = $this->categoryService->getById($request->id)->image;
-            unlink(public_path('img') . '/' . $oldImagePath);
+            if (file_exists(public_path('img') . '/' . $oldImagePath)) {
+                unlink(public_path('img') . '/' . $oldImagePath);
+            }
 
         }
         // Public Folder
